Give up battle with Texlive
We've been trying to minimie the size of the Docker image by
using tlmgr to load only those additional packages and fonts
that we actually need. When this works, the Docker image is only
1GB. But Debian's Texlive packages lag behiind the Texlive archives,
which breaks tlmgr for (part of) the year. So we give up and load
everything using the Debian package manager, yielding a 3.88GB Docker
image.